+++ bdsg/include/bdsg/ch.hpp
@@ -43,12 +43,84 @@ int edge_diff(ContractedGraph::vertex_descriptor nid, ContractedGraph& ch, CHOve
 void contract(CHOverlay::vertex_descriptor nid, ContractedGraph& ch, CHOverlay& ov, vector<DIST_UINT>& node_dists, vector<bool>& shouldnt_contract, int hop_limit);
 
 void make_contraction_hierarchy(CHOverlay& ov);
-
+ 
+template <typename ItrType>
+ItrType get_dist_itr(ItrType start_itr, ItrType hub_itr) {
+  auto node_count = *start_itr;
+  auto last_fwd_end_bound_itr = next(start_itr, 1+node_count);
+  if (hub_itr >= next(start_itr, *last_fwd_end_bound_itr)) {
+    //backwards label
+    auto first_back_bound_itr = next(start_itr, 1+node_count+1); 
+    auto last_back_bound_itr = next(start_itr, 1+node_count+1+node_count);
+    auto jump_to_dist = (*last_back_bound_itr) - *first_back_bound_itr;
+    return next(hub_itr, jump_to_dist);
+  } else {
+    //forwards label
+    auto first_fwd_bound_itr = next(start_itr, 1); 
+    auto last_fwd_bound_itr = next(start_itr, 1+node_count);
+    auto jump_to_dist = (*last_fwd_bound_itr) - *first_fwd_bound_itr;
+    return next(hub_itr, jump_to_dist); 
+  }
+} 
 
 DIST_UINT binary_intersection_ch(vector<HubRecord>& v1, vector<HubRecord>& v2);
+/*
+start_bound_index variables are relative to start_offset
+*/
+template <typename ItrType>
+DIST_UINT binary_intersection_ch(ItrType start_itr, size_t v1_start_bound_index, size_t v2_start_bound_index) {
+  auto v1_start_bound_itr = next(start_itr, v1_start_bound_index);
+  auto v1_end_bound_itr = next(v1_start_bound_itr, 1);
+  auto v2_start_bound_itr = next(start_itr, v2_start_bound_index);
+  auto v2_end_bound_itr = next(v2_start_bound_itr, 1);
+
+  auto v1_start_itr = next(start_itr, *v1_start_bound_itr); 
+  auto v1_end_itr = next(start_itr, *v1_end_bound_itr);
+  auto v2_start_itr = next(start_itr, *v2_start_bound_itr); 
+  auto v2_end_itr = next(start_itr, *v2_end_bound_itr); 
+  auto v1_range = ranges::subrange<ItrType>(v1_start_itr, v1_end_itr);
+  auto v2_range = ranges::subrange<ItrType>(v2_start_itr, v2_end_itr); 
+
+  auto& key_vec = v1_range.size() < v2_range.size() ? v1_range : v2_range; 
+  auto& search_vec = v1_range.size() < v2_range.size() ? v2_range : v1_range;
+ 
+  auto search_start_itr = search_vec.begin();
+  auto search_end_itr = search_vec.end(); 
+  DIST_UINT min_dist = INF_INT;
+  for (auto it = key_vec.begin(); it < key_vec.end(); it++) {
+    //cerr << "loop top" << endl;
+    auto k = *it;
+    auto k_dist_itr = get_dist_itr(start_itr, it); 
+    //cerr << "dist for k " << k << " is " << *k_dist_itr << ", at: " << distance(start_itr,k_dist_itr) << endl;
+    //cerr << "searching for " << k << " between " << distance(start_itr,search_start_itr) << " & " << distance(start_itr,search_end_itr) << endl;
+    search_start_itr = lower_bound(search_start_itr, search_end_itr, k); 
+    if (search_start_itr == search_end_itr) {
+      return min_dist;
+    } 
+    if (*search_start_itr == k) {
+      //cerr << "match found, key: " << *search_start_itr << ", at " << distance(start_itr,search_start_itr) << endl;
+      auto dist_itr = get_dist_itr(start_itr, search_start_itr);
+      DIST_UINT d = *(dist_itr) + *(k_dist_itr);
+      //cerr << "dist for key is: " << *dist_itr << ", at " << distance(start_itr,dist_itr) << endl; 
+      //cerr << "total dist is: " << d << endl;
+      min_dist = min(min_dist, d);
+    }
+  }  
+  return min_dist; 
+}  
 
 template <typename ItrType>
-DIST_UINT hhl_query(ItrType start_itr, size_t rank1, size_t rank2);  
+DIST_UINT hhl_query(ItrType start_itr, size_t rank1, size_t rank2) {
+  size_t label_count = *start_itr;
+
+  auto start_index_1 = 1+rank1;
+  auto start_index_2 = 1+label_count+1+rank2;
+  
+  DIST_UINT dist = binary_intersection_ch(start_itr, start_index_1, start_index_2);
+
+
+  return dist; 
+}      
 
 void down_dijk(int node, CHOverlay& ov, vector<DIST_UINT>& node_dists, vector<vector<HubRecord>>& labels, vector<vector<HubRecord>>& labels_rev);
